refactor: Factor date related functions out into a new file
The date related functions of viewUpdater have been extracted into a new
file dateViewUpdater.
